Refactored the DynamicTypeSupport header and source files
Moved DynamicSample and the dynamic DW and DR to their own files.
